+++ src/main/java/com/nixon/myreads/client/response/AuthorsResponse.java
@@ -1,4 +1,4 @@
 package com.nixon.myreads.client.response;
 
-public record AuthorsResponse(int id, String name) {
+public record AuthorsResponse(Long id, String name) {
 }
