+++ src/main/java/com/nixon/myreads/dto/response/AuthorsResponseDTO.java
@@ -1,4 +1,4 @@
 package com.nixon.myreads.dto.response;
 
-public record AuthorsResponseDTO(int id, String name) {
+public record AuthorsResponseDTO(Long id, String name) {
 }
